requirements.txt
Django==3.1
djangorestframework==3.11.1
django-extensions==2.2.9
psycopg2-binary==2.8.5
requests==2.24.0
requests-mock==1.8.0
django-model-utils==4.0.0
django-filter==2.3.0
django-qr-code==1.2.0
drf-yasg==1.17.1
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-mock==3.3.1
freezegun==1.0.0
marshmallow==3.7.1
django-post_office==3.4.1